Add User module to create, destroy and authenticate the user
For the time being this is just set up for a single user. The user's
email and hashed password are stored in the db/users.json file. Those
credentials are used to authenticate the user when they attempt to
login.

Eventually this will probably get moved to a database and support for
multiple users will be added. This is just something quick and easy to
get started with.